fai_m7_v3/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# SEC EDGAR API Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
# XBRL library for financial data parsing
python-xbrl==1.1.1

# Vector Database (Optional - for enhanced features)
pinecone-client==3.0.0
openai==1.6.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Logging
python-json-logger==2.0.7

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
